feat: decreasing loading time on these funding map pages
